add extract feature and modify code into parameterized function

+++ downloadscript.py
@@ -1,15 +1,39 @@
 import urllib.request
 import os
 import time
+import gzip
+
 
 number = "0001"
 file = "medline"
 
-for i in range(1,9):
-    number = "{:04}".format(i)
+#filenumber = input the file number on pubMed Database
+#this will download the file ./src/zip and extract to ./src/raw
+#func will return the unzipped file path
+def dlext(filenumber):
+    number = "{:04}".format(filenumber)
     file = "medline17n" + number + ".xml.gz"
     url = 'ftp://ftp.ncbi.nlm.nih.gov/pubmed/baseline/' + file
-    path = os.path.join("D:/INDIV/src/",file)
-    urllib.request.urlretrieve(url,path)
-    print("Downloading "+file)
+    if not os.path.exists('./src/zip'):
+            os.makedirs('./src/zip')
+    path = os.path.join("./src/zip",file)
+    
+    if os.path.exists(path):
+        print('Already got ' + file + ' proceeding to next step...')
+    else:
+        print("Downloading "+file)
+        urllib.request.urlretrieve(url,path)
+        print(file + " Downloaded")
+
+    print('unzipping ' + file)
+    with gzip.open(path, "rb") as zf:
+        unzipfile = "medline" + number + ".xml"
+        if not os.path.exists('./src/raw'):
+            os.makedirs('./src/raw')
+        unzippath = os.path.join('./src/raw', unzipfile)
+        with open(unzippath,"wb") as of:
+            of.write(zf.read())
+        print('unzipped ' + file)
+
     time.sleep(20)
+    return unzippath
